review: extract /api/anchor constants into sibling config.ts
Follows the route+config convention used by app/api/receipt/price and
app/api/verification — keeps route.ts focused on request handling and
makes the denylist / transient-error codes importable from tests.
